[#6922, #6923] Integrate quickstart dialog into welcome screen
Merges the adventure quickstart dialog into the welcome screen,
displaying the import options beneath the system settings.

Adds a new `dnd5e.ready` hook that is called only after and system
migrations have completed. This is when the welcome dialog is
displayed and gives modules the ability to avoid doing anything
that might disrupt system migrations.
