+++ module/applications/adventure/_module.mjs
@@ -1,2 +1,1 @@
 export {default as AdventureImporter5e} from "./importer.mjs";
-export {default as AdventureQuickstartDialog} from "./quickstart-dialog.mjs";
