+++ templates/apps/welcome-main.hbs
@@ -1,8 +1,18 @@
 <section class="tab {{~#if tab.active}} active{{/if}}" data-group="{{ tab.group }}" data-tab="{{ tab.id }}">
     <img class="hero" src="systems/dnd5e/ui/official/dnd5e-banner.webp">
     <div class="message">{{{ message }}}</div>
+
+    {{!-- Settings --}}
     <fieldset class="card">
         <legend>{{ localize "DND5E.WELCOME.Settings.Label" }}</legend>
         {{> "dnd5e.fieldlist" fields }}
     </fieldset>
+
+    {{!-- Adventure Import Actions --}}
+    {{#if importActions.length}}
+    <fieldset class="card">
+        <legend>{{ localize "DND5E.ADVENTURE.Options" }}</legend>
+        {{> "dnd5e.fieldlist" importActions }}
+    </fieldset>
+    {{/if}}
 </section>
